Update DBCDStorage with encryptedID changes

+++ DBFileReaderLib/DBParser.cs
@@ -110,13 +110,13 @@ namespace DBFileReaderLib
             return reader.GetEncryptedSections().Where(s => s != null).ToDictionary(s => s.TactKeyLookup, s => s.NumRecords);
         }
 
-        public Dictionary<int, int[]> GetEncryptedIDs()
+        public Dictionary<ulong, int[]> GetEncryptedIDs()
         {
             var reader = this._reader as IEncryptionSupportingReader;
 
             if (reader == null || reader.GetEncryptedIDs() == null)
             {
-                return new Dictionary<int, int[]>();
+                return new Dictionary<ulong, int[]>();
             }
 
             return reader.GetEncryptedIDs();
